# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShopIA API",
    "version": "1.0.0",
    "description": "Documentation API ShopIA Dashboard"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/api/boutiques/livraison/stats": {
      "get": {
        "summary": "Statistiques globales des services de livraison",
        "description": "Retourne les KPIs agrégés pour le suivi des entreprises de livraison.\n\n**Métriques** :\n- Partenaires actifs\n- Commandes du mois\n- Montant collecté\n- Taux de livraison\n",
        "tags": [
          "Livraison"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Statistiques récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "partenaires_actifs": {
                          "type": "integer",
                          "example": 100
                        },
                        "partenaires_actifs_tendance": {
                          "type": "number",
                          "example": 5
                        },
                        "commandes_mois": {
                          "type": "integer",
                          "example": 50
                        },
                        "commandes_mois_tendance": {
                          "type": "number",
                          "example": 8
                        },
                        "montant_collecte": {
                          "type": "number",
                          "example": 100000
                        },
                        "montant_collecte_tendance": {
                          "type": "number",
                          "example": 3
                        },
                        "taux_livraison": {
                          "type": "number",
                          "example": 88.5
                        },
                        "taux_livraison_tendance": {
                          "type": "number",
                          "example": 3
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/boutiques/livraison/incidents": {
      "get": {
        "summary": "Incidents de livraison - Commandes non livrées",
        "description": "Liste des commandes non livrées avec les motifs",
        "tags": [
          "Livraison"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Recherche par ID commande ou boutique"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Incidents récupérés avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id_commande": {
                                "type": "string"
                              },
                              "boutique_partenaire": {
                                "type": "string"
                              },
                              "livreur": {
                                "type": "string"
                              },
                              "motif": {
                                "type": "string"
                              },
                              "date": {
                                "type": "string",
                                "format": "date"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}